+++ include/npf/content/media.hpp
@@ -5,7 +5,7 @@
 #ifndef TUMBLRAPI_MEDIA_HPP
 #define TUMBLRAPI_MEDIA_HPP
 
-#include "content.hpp"
+#include "npf/content/content.hpp"
 
 /**
  * Media objects are used for image blocks, all kinds of posters (GIF, video, etc), native audio, native video,
