+++ include/npf/content/video.hpp
@@ -5,9 +5,9 @@
 #ifndef TUMBLRAPI_VIDEO_HPP
 #define TUMBLRAPI_VIDEO_HPP
 
-#include "content.hpp"
-#include "image.hpp"
-#include "attribution.hpp" // Circular dependency with image.hpp
+#include "npf/content/content.hpp"
+#include "npf/content/image.hpp"
+#include "npf/attribution.hpp" // Circular dependency with image.hpp
 
 /**
  * TODO Documentation
