+++ include/npf/layout/carousel.hpp
