+++ include/npf/layout/condensed.hpp
