+++ include/npf/trail.hpp
