+++ src/attribution.cpp
@@ -2,7 +2,7 @@
 // Created by Spud on 7/14/21.
 //
 
-#include "attribution.hpp"
+#include "npf/attribution.hpp"
 
 /*
 {
