+++ src/link.cpp
@@ -2,7 +2,7 @@
 // Created by Spud on 7/14/21.
 //
 
-#include "link.hpp"
+#include "npf/content/link.hpp"
 
 /*
 {
